Factor out ItemView base class from PlaylistView
